chore: mv lib/server/* lib/
